Cleanup unnecessary unboxing
Unboxing is unnecessary under Java 5 and newer, and can be safely removed.

+++ acl/src/main/java/org/springframework/security/acls/AclPermissionEvaluator.java
@@ -127,8 +127,7 @@ public class AclPermissionEvaluator implements PermissionEvaluator {
 
 	List<Permission> resolvePermission(Object permission) {
 		if (permission instanceof Integer) {
-			return Arrays.asList(permissionFactory.buildFromMask(((Integer) permission)
-					.intValue()));
+			return Arrays.asList(permissionFactory.buildFromMask((Integer) permission));
 		}
 
 		if (permission instanceof Permission) {
